Added HallOfFamePlayer, StaffMember and StaffMemberAvatarLayer domain entities and the related XmlFile entities, enums, contants, parsers, persisters, etc. Renamed columns and tables. Changed a LeagueCup CurrentRound type from uint to int since WorldDetails file values can be negative. Redone migrations to start clean. Cleaned and re-formatted code.

+++ src/Hyperar.HPA.Infrastructure/DataAccess/Mappings/LeagueCup.cs
@@ -42,7 +42,7 @@
                 .HasColumnName(Constants.ColumnName.CurrentRound)
                 .HasColumnOrder(
                     this.GetCurrentColumnOrder())
-                .HasColumnType(Constants.ColumnType.BigInt)
+                .HasColumnType(Constants.ColumnType.Int)
                 .IsRequired();
 
             builder.Property(p => p.RoundsLeft)
